add deregister task definition function
